Worked on Audio & collisons
- sound effect plays when the player shoots
- sound effect should play when bullet hits a target but doesn't
- worked on collisions w/ bullet & targets: bullet collides w/ multiple targets when going at a high speed. Shouldn't do that. WIP

+++ Assets/Final Project/Scripts/MainMenuNav.cs
@@ -23,9 +23,13 @@ public class MainMenuNav : MonoBehaviour
     {
         //reset game for level 1
         PlayerPrefs.SetInt("SCORE", 0);
-        PlayerPrefs.SetInt("ACCURACY", 0);
+        PlayerPrefs.SetFloat("ACCURACY", 0);
         PlayerPrefs.SetInt("SHOTS FIRED", 0);
         PlayerPrefs.SetInt("HIT SHOTS", 0);
+
+        //use for debugging or resetting the saved score
+        //PlayerPrefs.SetFloat("HIGH SCORE", 0);
+        
         //load first level
         SceneManager.LoadScene("Level_01");
     }
